Fix warnings in FactoryReport::init()

+++ OPHD/UI/Reports/FactoryReport.cpp
@@ -16,7 +16,7 @@
 
 using namespace NAS2D;
 
-static float SORT_BY_PRODUCT_POSITION = 0;
+static int SORT_BY_PRODUCT_POSITION = 0;
 
 static Rectangle<int> FACTORY_LISTBOX;
 static Rectangle<int> DETAIL_PANEL;
@@ -145,7 +145,7 @@ void FactoryReport::init()
 	btnShowDisabled.type(Button::Type::BUTTON_TOGGLE);
 	btnShowDisabled.click().connect(this, &FactoryReport::btnShowDisabledClicked);
 
-	float position_x = Utility<Renderer>::get().width() - 110;
+	int position_x = static_cast<int>(Utility<Renderer>::get().width()) - 110;
 	add(&btnIdle, position_x, 35);
 	btnIdle.type(Button::Type::BUTTON_TOGGLE);
 	btnIdle.size({140, 30});
